refactor: rename types file
