+++ src/libs/contentful/index.ts
@@ -1,3 +1,3 @@
-export * from './contentful'
+export * from './types'
 export { ExclusivityPane, BusinessPane, Offers, OffersWithCover, VenuesModule } from './moduleTypes'
 export { processHomepageEntry } from './processHomepageEntry'
